Got a lot of the front end stuff for the basic allowance card wired up correctly and calling the right APIs. I'm getting some code smells, need to take a good look at what I have, organize, and refactor, but this is a great start.

+++ Source/AllowanceApp.Shared/DTO/AllowancePointDTO.cs
@@ -8,8 +8,8 @@ namespace AllowanceApp.Shared.DTO
         public int AccountID { get; init; }
         public string Category { get; init; } = string.Empty;
         public int Price { get; init; }
-        public double Points { get; init; }
-        public double Total { get; init; }
+        public int Points { get; init; }
+        public int Total { get; init; }
 
         [JsonConstructor]
         public AllowancePointDTO() {}
